packages/scheduler_executor/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "wealth_management_portal_scheduler_executor" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/packages/scheduler_executor/html'
 --cov-report xml:'../../coverage/packages/scheduler_executor/coverage.xml'
 --html='../../reports/packages/scheduler_executor/unittests/html/index.html'
 --junitxml='../../reports/packages/scheduler_executor/unittests/junit.xml'
 --ignore=tests/integration
"""
markers = [
  "integration: marks tests as integration tests (require real AWS resources)"
]

[tool.hatch.build.targets.wheel]
packages = [ "wealth_management_portal_scheduler_executor" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 120
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[project]
name = "wealth_management_portal.scheduler_executor"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
  "aws-lambda-powertools==3.24.0",
  "aws-lambda-powertools[tracer]==3.24.0",
  "aws-lambda-powertools[parser]==3.24.0",
  "boto3>=1.34.0"
]

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "moto[dynamodb]>=5.0.0",
  "python-dotenv>=1.1.0",
]